backend/requirements.txt
django>=5.0
djangorestframework
celery
redis
psycopg2-binary
neo4j
qdrant-client
langchain
langchain-openai
openai
anthropic
google-genai
pymupdf
spacy
pdfplumber
python-dotenv
django-cors-headers
python-docx
dj-database-url
